Add WPF Application to the solution
Added a new WPF application to the solution with structure for MVVM pattern, necessary dependencies and settings. This WPF application will serve as the UI for our project. It includes a model for Person, a main window and a view model that supports adding person objects to an observable collection.